backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic-settings==2.5.2
sqlalchemy==2.0.36
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
# passlib 1.7.4 reads bcrypt.__about__; removed in bcrypt 4.1+
bcrypt==4.0.1
google-api-python-client==2.149.0
google-auth==2.35.0
